All ratings in ints now. avg to aPCG has revealed stdDev error.

+++ IPCG/src/Model/Component.java
@@ -8,8 +8,8 @@ public abstract class Component extends GameCollection<Part> {
 	}
 
 	@Override
-	public double rate() {
-		double rating = 0;
+	public int rate() {
+		int rating = 0;
 		for (Part element : elements) {
 			rating += element.rate();
 		}
